add: kokoro draft file
