+++ backend/models/stt/stt.py
