+++ backend/models/stt/whisper.py
